Use f64 for planet size
We use f64 for all other game numbers, so planet size should be
consistent.

+++ source/rust/vndf/src/client/graphics/draw/planet.rs
@@ -37,7 +37,7 @@ impl PlanetDrawer {
                 graphics: &mut Graphics,) {
         for (id, &(body, planet)) in &frame.planets {
             let transform = transforms.symbol_to_screen(cast(body.position));
-            let scale = planet.size * self.scaling_factor;
+            let scale = planet.size as f32 * self.scaling_factor;
 
             // draw selection behind planet
             if frame.select_ids.contains(id) {
